// src/app/@data/model/User/user.ts
export interface User {
  name?: string;
  lastName?: string;
  code?: number;
  password: string;
  estatus?: boolean;
  token?: string;
  personID?: number;
  email?: string;
  mobilPhone?: string;
  rolId?: number;
  userName?: string;
  typeDocument?: string;
  numTypeDocument?: string;
  alias?:string;
}